requirements.txt
# Core
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0

# Azure
azure-identity==1.14.0
azure-mgmt-resource==23.4.0
azure-mgmt-compute==31.0.0
azure-mgmt-storage==21.1.0
azure-mgmt-network==29.0.0

# LLM
ollama==0.1.6
sentence-transformers==2.5.1

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Templating
jinja2==3.1.2

# Utilitiesn
httpx==0.25.2

tenacity==8.2.3